# openapi.yaml
openapi: 3.0.3
info:
  title: LI Council API
  description: |
    API for LI Council mobile application and admin panel.

    ## Authentication
    This API uses Laravel Sanctum for authentication. Include the Bearer token in the Authorization header:
    `Authorization: Bearer {your-token-here}`

    ## Base URL
    - Production: `https://li-council.test/api`
    - Development: `http://localhost:8000/api`

    ## Documentation
    For detailed documentation with examples, see [API_DOCUMENTATION.md](./API_DOCUMENTATION.md)
  version: 1.0.0
  contact:
    name: LI Council API Support
    email: support@li-council.com
servers:
  - url: https://li-council.test/api
    description: Production server
  - url: http://localhost:8000/api
    description: Development server
security:
  - sanctumAuth: []
...
paths:
...
  /promoters:
    get:
      tags:
        - Promoters
      summary: Get promoters
      description: Get a paginated list of promoters
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Search'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Promoters retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedPromoters'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Promoters
      summary: Create promoter
      description: Create a new promoter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromoterRequest'
      responses:
        '201':
          description: Promoter created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Promoter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
components:
  securitySchemes:
    sanctumAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
    Page:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        default: 1
        minimum: 1
    PerPage:
      name: per_page
      in: query
      description: Number of items per page
      schema:
        type: integer
        default: 15
        minimum: 1
        maximum: 100
    Search:
      name: search
      in: query
      description: Search term
      schema:
        type: string
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
...
    Promoter:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        username:
          type: string
          example: "promoter123"
        email:
          type: string
          format: email
          example: "john@example.com"
        phone:
          type: string
          example: "+91-9876543210"
        state:
          type: string
          example: "Maharashtra"
        district:
          type: string
          example: "Mumbai"
        sub_district:
          type: string
          example: "Andheri"
        village:
          type: string
          example: "Andheri East"
        is_active:
          type: boolean
          example: true
        is_logged_in:
          type: boolean
          example: false
        profile_image:
          type: string
          nullable: true
          example: "storage/promoters/profile.jpg"
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreatePromoterRequest:
      type: object
      required:
        - name
        - username
        - email
        - password
        - state
        - district
        - sub_district
        - village
      properties:
        name:
          type: string
          example: "New Promoter"
        username:
          type: string
          example: "newpromoter"
        email:
          type: string
          format: email
          example: "promoter@example.com"
        phone:
          type: string
          example: "+91-9876543210"
        password:
          type: string
          example: "password123"
        state:
          type: string
          example: "Maharashtra"
        district:
          type: string
          example: "Mumbai"
        sub_district:
          type: string
          example: "Andheri"
        village:
          type: string
          example: "Andheri East"
        is_active:
          type: boolean
          default: true
    PaginatedPromoters:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Promoter'
        current_page:
          type: integer
          example: 1
        last_page:
          type: integer
          example: 5
        per_page:
          type: integer
          example: 15
        total:
          type: integer
          example: 75
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  message:
                    example: "Unauthenticated"
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...